+++ tests/CalendR/Test/Period/HourTest.php
@@ -2,6 +2,8 @@
 
 namespace CalendR\Test\Period;
 
+use CalendR\Period\Factory;
+use CalendR\Period\FactoryInterface;
 use CalendR\Period\Second;
 use CalendR\Period\Minute;
 use CalendR\Period\Hour;
@@ -11,14 +13,6 @@ use CalendR\Period\Year;
 
 class HourTest extends \PHPUnit_Framework_TestCase
 {
-
-    /**
-     * Data Provider: Invalid Constructors
-     *
-     * @static
-     * @access public
-     * @return array
-     */
     public static function providerConstructInvalid()
     {
         return array(
@@ -27,13 +21,6 @@ class HourTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * Data Provider: Valid Constructors
-     *
-     * @static
-     * @access public
-     * @return array
-     */
     public static function providerConstructValid()
     {
         return array(
@@ -44,15 +31,12 @@ class HourTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @param \DateTime $start
-     *
      * @dataProvider providerConstructInvalid
      * @expectedException \CalendR\Period\Exception\NotAnHour
      */
     public function testConstructInvalid($start)
     {
-        $calendar = new \CalendR\Calendar;
-        new Hour($start, $calendar->getFactory());
+        new Hour($start, $this->prophesize(FactoryInterface::class)->reveal());
     }
 
     /**
@@ -62,16 +46,9 @@ class HourTest extends \PHPUnit_Framework_TestCase
      */
     public function testConstructValid($start)
     {
-        new Hour($start);
+        new Hour($start, $this->prophesize(FactoryInterface::class)->reveal());
     }
 
-    /**
-     * Data Provider: Contains
-     *
-     * @static
-     * @access public
-     * @return array
-     */
     public static function providerContains()
     {
         return array(
@@ -94,18 +71,11 @@ class HourTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * Test: Contains
-     *
-     * @access public
-     * @param DateTime $start
-     * @param DateTime $contain
-     * @param DateTime $notContain
      * @dataProvider providerContains
-     * @return void
      */
     public function testContains($start, $contain, $notContain)
     {
-        $hour = new Hour($start);
+        $hour = new Hour($start, $this->prophesize(FactoryInterface::class)->reveal());
         $this->assertTrue($hour->contains($contain));
         $this->assertFalse($hour->contains($notContain));
     }
@@ -118,11 +88,11 @@ class HourTest extends \PHPUnit_Framework_TestCase
      */
     public function testGetNext()
     {
-        $hour = new Hour(new \DateTime('2012-01-01 05:00'));
+        $hour = new Hour(new \DateTime('2012-01-01 05:00'), $this->prophesize(FactoryInterface::class)->reveal());
         $this->assertEquals('2012-01-01 06:00', $hour->getNext()->getBegin()->format('Y-m-d H:i'));
-        $hour = new Hour(new \DateTime('2012-01-31 14:00'));
+        $hour = new Hour(new \DateTime('2012-01-31 14:00'), $this->prophesize(FactoryInterface::class)->reveal());
         $this->assertEquals('2012-01-31 15:00', $hour->getNext()->getBegin()->format('Y-m-d H:i'));
-        $hour = new Hour(new \DateTime('2013-02-28 23:00'));
+        $hour = new Hour(new \DateTime('2013-02-28 23:00'), $this->prophesize(FactoryInterface::class)->reveal());
         $this->assertEquals('2013-03-01 00:00', $hour->getNext()->getBegin()->format('Y-m-d H:i'));
     }
 
@@ -134,11 +104,11 @@ class HourTest extends \PHPUnit_Framework_TestCase
      */
     public function testGetPrevious()
     {
-        $hour = new Hour(new \DateTime('2012-01-01 00:00'));
+        $hour = new Hour(new \DateTime('2012-01-01 00:00'), $this->prophesize(FactoryInterface::class)->reveal());
         $this->assertEquals('2011-12-31 23:00', $hour->getPrevious()->getBegin()->format('Y-m-d H:i'));
-        $hour = new Hour(new \DateTime('2012-01-31 05:00'));
+        $hour = new Hour(new \DateTime('2012-01-31 05:00'), $this->prophesize(FactoryInterface::class)->reveal());
         $this->assertEquals('2012-01-31 04:00', $hour->getPrevious()->getBegin()->format('Y-m-d H:i'));
-        $hour = new Hour(new \DateTime('2012-01-31 06:00'));
+        $hour = new Hour(new \DateTime('2012-01-31 06:00'), $this->prophesize(FactoryInterface::class)->reveal());
         $this->assertEquals('2012-01-31 05:00', $hour->getPrevious()->getBegin()->format('Y-m-d H:i'));
     }
 
@@ -150,7 +120,7 @@ class HourTest extends \PHPUnit_Framework_TestCase
      */
     public function testGetDatePeriod()
     {
-        $hour = new Hour(new \DateTime('2012-01-31 13:00'));
+        $hour = new Hour(new \DateTime('2012-01-31 13:00'), $this->prophesize(FactoryInterface::class)->reveal());
         $i = 0;
         foreach ($hour->getDatePeriod() as $dateTime) {
             $i++;
@@ -170,7 +140,7 @@ class HourTest extends \PHPUnit_Framework_TestCase
         $currentDateTime = new \DateTime();
         $otherDateTime = clone $currentDateTime;
         $otherDateTime->add(new \DateInterval('PT5H'));
-        $currentHour = new Hour(new \DateTime(date('Y-m-d H:00')));
+        $currentHour = new Hour(new \DateTime(date('Y-m-d H:00')), $this->prophesize(FactoryInterface::class)->reveal());
         $otherHour = $currentHour->getNext();
         $this->assertTrue($currentHour->contains($currentDateTime));
         $this->assertFalse($currentHour->contains($otherDateTime));
@@ -185,8 +155,8 @@ class HourTest extends \PHPUnit_Framework_TestCase
      */
     public function testToString()
     {
-        $hour = new Hour(new \DateTime(date('Y-m-d H:00')));
-        $this->assertSame($hour->getBegin()->format('G'), (string) $hour);
+        $hour = new Hour(new \DateTime(date('Y-m-d H:00')), $this->prophesize(FactoryInterface::class)->reveal());
+        $this->assertSame($hour->getBegin()->format('G'), (string)$hour);
     }
 
     /**
@@ -204,69 +174,44 @@ class HourTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * Test: Includes
-     *
-     * @access public
-     * @param DateTime $begin
-     * @param PeriodInterface $period
-     * @param boolean $strict
-     * @param boolean $result
      * @dataProvider includesDataProvider
-     * @return void
      */
     public function testIncludes(\DateTime $begin, PeriodInterface $period, $strict, $result)
     {
-        $hour = new Hour($begin);
+        $hour = new Hour($begin, $this->prophesize(FactoryInterface::class)->reveal());
         $this->assertSame($result, $hour->includes($period, $strict));
     }
 
-    /**
-     * Test: Format
-     *
-     * @access public
-     * @return void
-     */
     public function testFormat()
     {
-        $hour = new Hour(new \DateTime(date('Y-m-d H:00')));
+        $hour = new Hour(new \DateTime(date('Y-m-d H:00')), $this->prophesize(FactoryInterface::class)->reveal());
         $this->assertSame(date('Y-m-d H:00'), $hour->format('Y-m-d H:i'));
     }
 
-    /**
-     * Test: Is Current?
-     *
-     * @access public
-     * @return void
-     */
     public function testIsCurrent()
     {
-        $currentHour = new Hour(new \DateTime(date('Y-m-d H:00')));
-        $otherHour   = new Hour(new \DateTime('1988-11-12 16:00'));
+        $currentHour = new Hour(new \DateTime(date('Y-m-d H:00')), $this->prophesize(FactoryInterface::class)->reveal());
+        $otherHour = new Hour(new \DateTime('1988-11-12 16:00'), $this->prophesize(FactoryInterface::class)->reveal());
         $this->assertTrue($currentHour->isCurrent());
         $this->assertFalse($otherHour->isCurrent());
     }
 
-    /**
-     * Data Provider: Includes
-     *
-     * @static
-     * @access public
-     * @return array
-     */
     public function includesDataProvider()
     {
+        $factory = $this->prophesize(FactoryInterface::class)->reveal();
+
         return array(
-            array(new \DateTime('2013-09-01 12:00'),  new Year(new \DateTime('2013-01-01')),            true,   false),
-            array(new \DateTime('2013-09-01 12:00'),  new Year(new \DateTime('2013-01-01')),            false,  true),
-            array(new \DateTime('2013-09-01 12:00'),  new Year(new \DateTime('2013-01-01')),            false,  true),
-            array(new \DateTime('2013-09-01 12:00'),  new Day(new \DateTime('2013-09-01')),             true,   false),
-            array(new \DateTime('2013-09-01 12:00'),  new Day(new \DateTime('2013-09-01')),             false,  true),
-            array(new \DateTime('2013-09-01 12:00'),  new Hour(new \DateTime('2013-09-01 12:00')),      true,   true),
-            array(new \DateTime('2013-09-01 12:00'),  new Hour(new \DateTime('2013-09-01 12:00')),      false,  true),
-            array(new \DateTime('2013-09-01 12:00'),  new Minute(new \DateTime('2013-09-01 12:34')),    true,   true),
-            array(new \DateTime('2013-09-01 12:00'),  new Minute(new \DateTime('2013-09-01 12:34')),    false,  true),
-            array(new \DateTime('2013-09-01 12:00'),  new Second(new \DateTime('2013-09-01 12:34:45')), false,  true),
-            array(new \DateTime('2013-09-01 12:00'),  new Second(new \DateTime('2013-09-01 12:34:45')), false,  true),
+            array(new \DateTime('2013-09-01 12:00'), new Year(new \DateTime('2013-01-01'), $factory), true, false),
+            array(new \DateTime('2013-09-01 12:00'), new Year(new \DateTime('2013-01-01'), $factory), false, true),
+            array(new \DateTime('2013-09-01 12:00'), new Year(new \DateTime('2013-01-01'), $factory), false, true),
+            array(new \DateTime('2013-09-01 12:00'), new Day(new \DateTime('2013-09-01'), $factory), true, false),
+            array(new \DateTime('2013-09-01 12:00'), new Day(new \DateTime('2013-09-01'), $factory), false, true),
+            array(new \DateTime('2013-09-01 12:00'), new Hour(new \DateTime('2013-09-01 12:00'), $factory), true, true),
+            array(new \DateTime('2013-09-01 12:00'), new Hour(new \DateTime('2013-09-01 12:00'), $factory), false, true),
+            array(new \DateTime('2013-09-01 12:00'), new Minute(new \DateTime('2013-09-01 12:34'), $factory), true, true),
+            array(new \DateTime('2013-09-01 12:00'), new Minute(new \DateTime('2013-09-01 12:34'), $factory), false, true),
+            array(new \DateTime('2013-09-01 12:00'), new Second(new \DateTime('2013-09-01 12:34:45'), $factory), false, true),
+            array(new \DateTime('2013-09-01 12:00'), new Second(new \DateTime('2013-09-01 12:34:45'), $factory), false, true),
         );
     }
 
@@ -279,7 +224,7 @@ class HourTest extends \PHPUnit_Framework_TestCase
     public function testIteration()
     {
         $start = new \DateTime('2012-01-15 13:00');
-        $hour = new Hour($start);
+        $hour = new Hour($start, new Factory());
         $i = 0;
         foreach ($hour as $minuteKey => $minute) {
             $this->assertTrue(is_int($minuteKey) && $minuteKey >= 0 && $minuteKey < 60);