+++ CHANGELOG.md
@@ -7,3 +7,4 @@ CHANGELOG
  * Non-strict mode does not exists anymore (use the factory / calendar to create periods)
  * No more google calendar provider (hard to maintain and a bit out of the scope of this library)
  * Removed first-monday / last-sunday in favor of first weekday / last weekday
+ * Factory argument is now mandatory for period instantiation